feat(secrets): Use `validation_state` for priority, not uniqueness (#9349)
With the addition of `anywhere` as a pattern key, we expect noise to
increase if we report all invalid matches in certain contexts. For
instance, a file with a secret of the form A + B (where we match on A,
and include B via `anywhere`) will result in one valid match, but if the
same file contains many such pairs we would get N valid matches and N^2
- N invalid matches (reduced post-deduplication to N invalid matches).
However, all such invalid matches are ultimately noise, since for A1,
..., AN we only fundamentally care about the valid match (the one
correspoding to (A1, B1), etc...).

To reduce noise, when two matches are considered equal under
`cli_unique_key` we now add a mechanism for replacing "lower priority"
matches with "higher priority" ones. For now, a finding is just higher
priority if it is confirmed valid and the other finding is not.

This removes validation_state from `cli_unique_key` since two findings
will be considered substatively equal at the same validation state, just
able to be higher priority within that equivalence class.

Test plan: See https://github.com/semgrep/semgrep-proprietary/pull/1137