chore!(rust): Update the semgrep-rust submodule (3/n) (#9390)
This updates past a major change to how Rust macros are handled in the
parser, as well as pulling in several smaller changes. I tried to keep
behavior similar but it was not always practical. I worked through a few
test failures related to macro matching. Fortunately the desired
behavior seemed fairly well-tested, but it's always possible with this
sort of change that some untested edge case no longer behaves the way
users expect. We'll just have to fix forward if such cases crop up.

This is a companion to
https://github.com/semgrep/ocaml-tree-sitter-semgrep/pull/463

Test plan: Automated tests