logging: Reduce noise in -debug's output
Lacking range info for an expression is pretty common unfortunately, due
to the use of fake tokens. Logging these adds a lot of noise and it is
unlikely to help us diagnosing any of the main issues that our customers
typically complain about.

test plan:
`semgrep -c p/default-v2 path/to/repo --debug` is less noisy