pre-commit: Pin Black to version 23.10.1
PR #9384 bumped Black version to 23.11.0 (from 22.6.0) which resulted in
an "InvalidManifestError" for some of us. Some appear to have fixed it
by doing a `pre-commit clean` and upgrading pre-commit to 3.6.0, but for
other these steps didn't fix it.

Everything seems to work fine with 23.10.1, which is still quite new,
so let's use that.

Fixes: 5402b1738fc ("chore(internal): Update pre-commit config (#9384)")

test plan:
Run pre-commit