fix(SSC): Fix newline bug in project.toml parser (#9411)
Previously the parser would fail on empty toml tables in
`pyproject.toml` files, because we'd consume all the whitespace when
parsing the empty table, and then there'd be no whitespace for the
`.sep_by(new_lines)` (which separates each table) to read. So I made the
separator new lines optional. Test included.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>