feat(metrics): add a ppx for tracing
Adding profiling is fairly difficult right now. We carry it through
layers of functions, adding to the profiling object as we go. This
is cumbersome and doesn't conform to standard practices. In particular,
there are standard reporting formats that we could take advantage of.

[ocaml-trace](https://github.com/c-cube/ocaml-trace) is a frontend for metrics backends that we can use
to help us with instrumentation. Using `ocaml-trace`, we can swap
different metrics backends depending on our needs. It is compatible
with OpenTelemetry, a common standard.

To make instrumentation easier, we would like to be able to instrument
functions using PPX annotations. This PR starts that work by creating
a PPX for tracing.

Test plan: (TODO)

Also TODO: will make it possible to specify your own action name