feat(metrics): add a ppx for tracing (#9448)
Adding profiling is fairly difficult right now. We carry it through
layers of functions, adding to the profiling object as we go. This is
cumbersome and doesn't conform to standard practices. In particular,
there are standard reporting formats that we could take advantage of.

[ocaml-trace](https://github.com/c-cube/ocaml-trace) is a frontend for
metrics backends that we can use to help us with instrumentation. Using
`ocaml-trace`, we can swap different metrics backends depending on our
needs. It is compatible with OpenTelemetry, a common standard.

To make instrumentation easier, we would like to be able to instrument
functions using PPX annotations. This PR starts that work by creating a
PPX for tracing.

Test plan: 

To see the result of transforming with the ppx run `dune describe` on
one of the files in `libs/tracing/tests`:
```
➜  ppx git:(emma/ea-305-create-ppx-for-profiling-using-ocaml-trace) ✗ pwd
/Users/emma/workspace/semgrep/libs/tracing/ppx
➜  ppx git:(emma/ea-305-create-ppx-for-profiling-using-ocaml-trace) ✗ dune describe pp libs/tracing/tests/basic.ml
Entering directory '/Users/emma/workspace/semgrep'
let foo a b = a + b [@@trace "foo 1"]

let foo a b =
  Trace_core.with_span ~__FILE__ ~__LINE__ "get_rules" @@ fun _sp -> a + b
```

To see trace output an actual trace locally, it takes a few steps:
1. Modify `Core_scan` as shown in the bottom comment of
`libs/tracing/Tracing.ml`. In particular, extract out the
`rules_from_rules_source` call into a function and annotate it with
`[@@trace]`, then call that function using the backend setup
2. Run:
```
➜  rules git:(emma/ea-305-create-ppx-for-profiling-using-ocaml-trace) ✗ pwd
/Users/emma/workspace/semgrep/tests/rules
➜  rules git:(emma/ea-305-create-ppx-for-profiling-using-ocaml-trace) ✗ sc -rules taint_array.yaml taint_array.py . -l python
```

This will produce a file, `trace.json`, which looks like:
```
[{"pid":28575,"cat":"","tid": 0,"dur": 125.75,"ts": 44962.58,"name":"Core_scan.get_rules","ph":"X"}]
```

Later, we'll make the instrumentation actually work and probably change
the backend! But this PR just adds the ppx for convenience