feat(LS): cache file targeting (#9447)
Currently we traverse the entire project on every scan, to determine if
a file or files are valid targets (i.e. not gitignore'd etc.). This can
be slow on large repos (like semgrep). This PR makes it so we cache the
possible targets, and only update the cache when a user
deletes/renames/creates a file, or updates what workspaces are being
used.

## Test plan
I confirmed manually, and that in the LS e2e tests these code paths are
flexed