fix: incorrect example rule does not cover all cases
In the example "Verify transactions before making them" rule, there is a method that should semantically be caught by the rule, but it's not. As this is one of the prominently featured in the README, this may lead to confusion when learning semgrep.

The issue is that the example file contains a method in which two distinct transactions are verified and committed respectively. However, the "not-pattern" ignores the parameter and thus, does not fail on the method, despite the example stating it would.