fix(lspjs): Final round of lspjs fixes (#9513)
A few more bug fixes for:
* Logging out
* Parsing large json scan configs

Yojson uses heavy tail recursion to do parsing (duh), but JS does NOT do
any TCOs. So while parsing large json blobs, it would get to 4gb+ of
memory, and then crash. To fix this I made it so we can inject the
native JS JSON parser, which doesn't crash.

## Test plan
Tested manually

See #team-pdx channel for a gif of this working