Use a function for line, col -> bytepos conversion
Currently we require a hashtable to be used but this is more restrictive
than necessary, and we've converted some code which was formerly
generating hashtables to now generate functions which serve the same
purpose. To be able to use those functions (now in Pos.ml), let's open
this type up a bit. Existing places where we use a Hashtbl and find are
now replaced with a function which calls find on the relevant Hashtbl.