feat: Use function for tree sitter parser line,col -> bytepos conversion (#9542)
Currently we require a hashtable to be used but this is more restrictive than
necessary, and we've converted some code which was formerly generating
hashtables to now generate functions which serve the same purpose. See, e.g.,
<https://github.com/semgrep/semgrep/pull/7596>,
<https://github.com/semgrep/semgrep/pull/9223>, (eventually replacing
`Parse_info.full_charpos_to_pos_large` with `Pos.full_charpos_to_pos_large`).
To be able to use those functions, let's open this type up a bit. Existing
places where we use a Hashtbl and find are now replaced with a function which
calls find on the relevant Hashtbl.