Initial support for tracing via Opentelemetry
Add two simple traces that can be activated via `--trace`. When
`--trace` isn't passed, the trace calls should be virtually no-ops.
When it is, spans for the trace will be sent to the collector.

Currently, this collector is just `localhost:4317`. Later, we'll
make it a permanent endpoint and include traces in metrics.

Test plan:
See instructions in notion "How to add tracing" to get this running

TODO: currently trace doesn't post anywhere except localhost. Should
I prepend `semgrep-internal` to it anyway?

TODO: does not yet work with multiple processes

TODO: instrument with the official ppx instead

TODO: add tags to the top level span for categorization

TODO: try the cottrp library for instrumentation