AST_to_IL: C++: Better translation of `delete` expressions
Instead of encoding `delete` as a string in Generic, we now prefer an
`OtherStmt` + `OS_Delete`. We also added a proper constructor for
`delete` in IL. Previously `delete` was encoded into IL as a `FixmeStmt`,
but it worked in many cases because `AST_to_IL` looks into the `any list`
and tries to translate the `any`s that correspond to expressions. But it
did not work for `at-exit: true` sinks, where these `FixmeStmt`s are
being ignored.

Helps PA-3339

test plan:
semgrep-core -lang cpp -dump_ast
TODO