osemgrep sarif output, part 1 (#9608)
This passes already two test cases :)

There are some issues, though:
- we are not using the opam package sarif yet (was easier to get started
this way)
- we couldn't find test cases for "security-severity" being present
- neither for "executionNotifications" being non-empty

The version in this PR is hardcoded to 1.56. We set the "hide_nudge" to
true in all cases.

When testing, we observed that osemgrep always applies autofixes (i.e.
the dryrun is set to false in src/fixing/Autofix.ml).

The failure behaviour if the json metadata is expected to be a string,
but some other json type is found, is unclear to us. We implemented a
"best_effort_string" which calls JSON.json_to_string in the latter case
(should we raise an exception instead?).

Other sarif test cases have mismatches on the fingerprints - not sure
whether the matchBasedId is expected to be identical between python and
OCaml.

And finally, we put the `fixed_lines` in Sarif_output.ml (instead of
putting it into the extra field of the cli_match in Cli_json_output),
since otherwise some (autofix) test cases failed (since they expect
`fixed_lines` to be not present). Any idea in which cases the
fixed_lines should be Some?

---------

Co-authored-by: Reynir <reynir@reynir.dk>