osemgrep fingerprints: escape single quotes (#9620)
Python escapes single quotes so we need to escape them as well when
computing the fingerprint. This fixes the osemgrep fingerprints for
rules with patterns that use single quotes, and made another sarif test
pass (not yet marked as such).

There are still issues with how fingerprints are computed. I ran both
pysemgrep and osemgrep patched so they print the fingerprint before
hashing and found the above error. There are still discrepancies when
there are multiple patterns. It seems the order is not always the same,
and in some cases pysemgrep seems to omit patterns(?!).