test: Fix broken pattern used in test (#9614)
In #9611, I am undoing a change that introduced lazy pattern parsing.
This restores correct error handling when invalid patterns are
encountered. This pattern is incorrect: It uses single quotes for a
string literal in Go. Go uses double quotes. Additionally, the single
quotes were escaped unnecessarily, which may have also caused issues.

Combined with my changes in #9611, this leads to a test failure. I guess
invalid rules aren't included in the list of rules stored for an LS
session, and there is an assertion in the `ci_tests` function in
`Unit_LS.ml` that checks the length of the list of rules. Without my
changes in #9611, this rule is incorrectly considered valid because the
pattern parse error does not appear at rule parsing time.

Test plan: Rebase #9611 on top of this and verify that `make core-test`
passes.