fix: Undo lazy pattern parsing
We have a bunch of error handling at rule creation time which nicely
handles pattern parse errors. An optimization made pattern parsing lazy,
so any exceptions related to pattern parsing are no longer raised at
rule creation time. Instead, they are raised whenever the patterns are
accessed downstream, in code that was not designed to handle errors with
patterns. Unfortunately, this means that a pattern parse error will mean
that any target to which that rule is applied will end up with no scan
results. This means, for example, that running a Python rule with an
invalid pattern will mean that semgrep-core reports no results at all
for Python files.

Now that we've addressed the performance issues with the Ruby parser,
parsing rule patterns up front is no longer unacceptably slow. It would
be nice to do them lazily, but without some investment in error
handling, it's too dangerous.

Partially addresses PA-3325.

Test plan:

Created a directory with just a single Java file. Ran `semgrep scan
--config p/default -d` to get the underlying `semgrep-core` command. Ran
that `semgrep-core` command 10 times before this change and 10 times
after.

The mean before was 1.233 s, the mean after was 1.891 s, with fairly low
standard deviations for both. This is an increase of about 650 ms. This
seems to be an acceptable performance hit.

Also, see the snapshot updates. The message changed slightly to indicate
that the error was handled during rule parsing. You can also see that
the scans completed, and that the list of scanned files is now nonempty.

Also, create a Semgrep org with only two rules in the policy: One python
rule with a syntax error in the pattern, and one that just looks for
`print("hi")`. Run `semgrep ci --oss-only`. Before, only the pattern
parse error is reported. After, both the pattern parse error and the
finding are reported. However, it still does not upload the findings to
the app.

Run the same test with `semgrep ci --pro`. Before, it errored out and
did not produce results. Now, it completes successfully, uploads results
to the app, and does not report the problem with the rule.

Future work should make both OSS and Pro report the error on the CLI but
also upload findings to the app.