fix: tainting: Precise identification of sinks
When computing the top matches for a sink (or a source or a sanitizer),
it is not enough to just look at the top origs of a node. For example,
this `{__html: props ? DOMPurify.sanitize(props.text) : ''}` will be
translated into (roughly) the following IL:

    if props:
      tmp := DOMPurify.sanitize(props.text);
    else:
      tmp := ''
    {__html: !tmp}

The expression whose `eorig` corresponds to the actual sink we want to
infer is `!tmp` and it is nested within a record expression. We have no
choice but to visit all the IL expressions and check every `eorig`.

Follows: dc397cc558d ("tainting: Make identification of sinks (even) more precise (#9342)")

Helps RULES-6457

test plan:
make test # new tests