fix: properly process `nosemgrep` with multiple ids (#9498)
**## What:
This PR makes it so that Semgrep doesn't err upon processing a
`nosemgrep` with multiple IDs.

## Why:
This is intended behavior, and should work.

## How:
We were not properly splitting the the list of IDs obtained by a
`nosemgrep` by comma, so this PR just re-introduces that behavior.

This causes a test to fail, due to some existing behavior of how
`nosemgrep` works. Essentially, for every match that is produced, if
there exists `#nosemgrep` comment for a rule which is not the same as
the match's ID, Semgrep will complain.

In the case where there are multiple `# nosemgrep` IDs for a single
line, this will always cause an error. This seems undesirable -- we
would prefer to check that each ID that `# nosemgrep` ignores truly has
a match it is ignoring. This would be considerably more work, however,
so this PR just disables the warning when there are multiple IDs being
ignored at once.

## Test plan:
Added a regression test.

Closes #9463 
Closes PA-3320**