osemgrep: implement junit_xml output (#9626)
This implements the junit_xml output for osemgrep. The single junit_xml
test does not pass because the order of the test results is different.
Otherwise it closely mimics the pysemgrep junit_xml output.

The reason for the difference in order is because the osemgrep code uses
the cli_output (json) which is sorted. Maybe the solution is to sort the
output in pysemgrep as well? Especially with the following comment in
mind:

    # Sort results so as to guarantee the same results across different
    # runs. Results may arrive in a different order due to parallelism
    # (-j option).