pysemgrep: sort rule matches in junit_xml output (#9636)
The motivation is because the output is sorted in e.g. the json output
for reproducibility in the presence of parallelism. Another motivation
is because osemgrep sorts the output and this change will align
pysemgrep with osemgrep and make the junit_xml test pass for osemgrep
too.

Previous discussion in https://github.com/semgrep/semgrep/pull/9626.