fix: tainting: Precise identification of sources/sanitizers/sinks (#9623)
When computing the best matches for a sink (or a source or a sanitizer),
it is not enough to just look at the top origs of each CFG node (... and
we knew it was not enough but it was left as a TODO). For example, this
`{__html: props ? DOMPurify.sanitize(props.text) : ''}` will be
translated into (roughly) the following IL:

    if props:
      tmp := DOMPurify.sanitize(props.text);
    else:
      tmp := ''
    {__html: !tmp}

The expression whose `eorig` corresponds to the actual sink we want to
infer is `!tmp` and it is nested within a record expression. Previously
we just looked at the "top" expression of each CFG node, so we looked at
`{__html: !tmp}` alone and we missed the nested `!tmp`.

Now we visit all the IL expressions and check every `eorig` so that we
don't miss anything.

Follows: dc397cc558d ("tainting: Make identification of sinks (even)
more precise (#9342)")

Helps RULES-6457

test plan:
make test # new tests