fix: report token errors properly
Investigation of https://github.com/semgrep/semgrep/issues/9628

When we can't figure out how to convert a row/column to a bytepos, we just put in
-1 and move on. In the linked Github issue, this causes an opaque exception when
we try to use the bytepos in autofix. Better to raise an exception.

Fixing this causes a pattern parse error in one of our Dockerfile rules. TODO that
parse error needs to be fixed as well in this PR.

Test plan: run p/default against Dockerfile in linked ticket
TODO automated tests