Osemgrep fingerprint (#9655)
This aligns how fingerprints are computed with pysemgrep.

First, `pattern-sanitizers` is omitted. This is done in pysemgrep due to
`"pattern-sanitizers"` is not in `RuleValidation.PATTERN_KEYS`.

https://github.com/semgrep/semgrep/blob/53caa62f16ed9e585f74d583dcce9a438dbe14b6/cli/src/semgrep/rule.py#L307

https://github.com/semgrep/semgrep/blob/53caa62f16ed9e585f74d583dcce9a438dbe14b6/cli/src/semgrep/rule_lang.py#L459-L472

Second, the sort order of the patterns was incorrect. It is not enough
to flatten the patterns and then sort them.
```yaml
- patterns:
  - pattern-either:
    - pattern: a
    - pattern: c
  - pattern: b
```
The above patterns should result in `a c b` as `a c` comes before `b`.
However, flattening the patterns we get `a b c`.

The first item seems to me like a bug in pysemgrep(?), but fixing it
makes a new test pass. The second item fixes the output but does not
make any new tests pass as the sort order is still different.

---------

Co-authored-by: ajbt200128 <austin@cutedogs.org>