feat(osemgrep): gracefully handle fatal rules (#9641)
## What:
This PR makes it so `osemgrep` can have graceful handling of fatal rule
parsing errors.

## Why:
We have a notion of recoverable and unrecoverable rule parsing errors.

A recoverable rule error is one such as an invalid rule ID, or invalid
rule language, or invalid pattern, dictated by
`Rule.invalid_rule_error`. These are packaged into the return type of
`rules_and_origins`, when calling
`Rule_fetching.rules_from_rules_source`.

Separately, we have unrecoverable rule errors. This can take the form of
invalid YAML, for instance.

The current behavior is that when calling
`Rule_fetching.rules_from_rules_source` on an invalid YAML rule, it will
raise the parsing exception, since these errors cannot fit into the type
of `Rule.invalid_rule_error` (and thus `Rule_fetching.rules_and_origin`,
which is only recoverable errors). This percolates the exception to the
top-level of `osemgrep`, and gives a default fatal exit code.

It is undesirable for us to let an exception just trickle to the
top-level of the OCaml run, when they are usually contained and reported
by `pysemgrep`. In particular, for invalid rule parsing, `pysemgrep`
gives an exit code of `7`, not the default fatal exit code. To maintain
consistency, we need to be able to recover from a rule parsing error in
`osemgrep`.

## How:
This PR makes it so that the `Rule_fetching.rules_from_dashdash_config`
primitive also returns a list of `Rule.error`s, which indicate the fatal
errors that can result from trying to obtain the rules. This means that
consumers now have the option of handling the errors.

In all the cases where we use the similar function
`Rule_fetching.rules_from_rules_source`, the errors are consumed and
potentially raise the `Semgrep_error` exception, which continues to our
established error-handling logic, following the `pysemgrep` behavior.
This is because in all the use sites of `rules_from_rules_source`, we
would not like to continue when we receive fatal rule parsing. As such,
it also does not return the `Rule.error list`, since it performs error
handling.

## Test plan:
Some tests now work properly with `osemgrep`. I also added a test
demonstrating the new error handling with `semgrep test`.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>