Add `source` field to `Xtarget.t`
This allows us to record a user-facing source for where the target came
from. Currently this is just a file, but could be expanded in the
future. Right now, it enables cases where we might have a tempfile but
want to report something at a different file for the user. In the
future, this could be used to report, e.g., a git object, network
request, more granular treatment of nested content or something else as
the source for content.