Add a GitBlob source
This allows (in principle, but not yet in the actual output) reporting
to the user that the source of a finding was from a git blob object.
This is useful for e.g., historical scans for secrets.

Currently this has a sha (of the blob object itself) and a list of
commits and the path to the blob at that commit.