Use Target_location.t in target_source
Formerly we used In.target here, but this is more restrictive than we
would like in tracking source info and is tied to the external json
input format. This requires some additional translation, but we were
already shuffling data around in most instances, so this just changes
lambdas for some maps.