Osemgrep fingerprint: omit pattern-propagators (#9676)
Like in https://github.com/semgrep/semgrep/pull/9655 I noticed
`"pattern-propagators"` is missing in `RuleValidation.PATTERN_KEYS` in
pysemgrep and so it occurs to me we should omit all `propagators`
formulae as well when computing the matchBasedId. Maybe @ajbt200128 can
confirm and has an explanation for this? There is no test case covering
this so there is no change in number of tests passing due to this
change.

This PR also marks a test osempass due to both #9655 and #9656 being
merged.