feat: Supply Chain in semgrep-core (#9695)
This is an initial pass at porting semgrep supply chain functionality
into semgrep-core. This PR should actually do *nothing*, it just sets up
a bunch of infrastructure. Here's a breakdown of the changes:

## Rule

Rules can now contain dependency patterns. See `Rule.ml`.

## Input_to_core

A target can now be annotated with a lockfile target, or be a lockfile
itself. See `Input_to_core.atd`.
Pysemgrep and osemgrep still only send regular code targets, with no
lockfiles, so nothing changes in terms of observable behavior.

## PatternMatch

Analogously to the situation with targets, a pattern match can now have
an optional dependency match attached to it, or itself be a
"lockfile-only" match. See `Pattern_match.ml`

## Supply Chain Types

We have some new types to describe the data of a parsed dependency and a
parsed dependency pattern. These do not make use of `Generic_ast`
because the data is much more constrained in shape.

## Core scan

The main scan entry point in `Core_scan.ml` now receives a list of
targets that are either `LockfileTargets` or `CodeTargets`. When
building the target handler (in `mk_target_handler`), we case on what
kind of target we have.
* If it's a `CodeTarget`, we proceed as normal, but with these
additions:
* For each rule we check if the rule has a dependency pattern and the
target has an associated lockfile target
* If both are present, we compute the set of dependencies from the
lockfile target that match the dependency patterns from the rule, and we
store this in a hash table by `rule_id`
* When generating code matches for a rule, we look up if it had any
associated dependency matches, and if so we annotate the code matches
with these dependency matches.
* There is some annoying complexity in this: If there was one dependency
match, then we just annotate each code match with it. But if there were
multiple (say *n*) dependency matches and a code match, there are two
possibilities:
1. There are *n* transitive copies of the dependency (that match), no
direct copy, and we are using it directly in code. There's no way to
tell which of the copies gets pulled in (this would be package-manager
dependent and probably even project dependent), so we generate a new
code match (annotated with the dependency match) for each possibility
2. There are *n - 1* transitive copies and one direct copy. The code
must be using the direct one, so we should only produce a single
annotated code match
* If it's a `LockfileTarget`, and the rule has a dependency pattern, we
generate a "lockfile-only" `Pattern_match.t` for each dependency in the
lockfile that matches the dependency pattern.
* The extra complication here is that if a rule with both a code pattern
*and* a dependency pattern produced a "reachable" (code + dependency)
match from a given lockfile, that rule should *not* produce a
lockfile-only match on that matched dependency. This deduping is not
currently implemented in this PR, though there is a comment about it in
`Core_scan.ml`

## Parsing

This PR contains *no parsing code*. All the of the cases were parsing is
required just return a dummy value. This is in service of keeping the PR
a little bit smaller. We would also like to put the lockfile parsers in
Semgrep Pro. PRs for this will follow this one.

## Testing

This has no tests because it does not introduce new usable functionality
yet. Tests will be included in future PRs to semgrep-pro that have
parsers.