chore(c/cpp): bring up to speed again (#9688)
## What:
This PR updates both the `semgrep-c` and `semgrep-cpp` modules with the
latest changes from `tree-sitter-c` and `tree-sitter-cpp`.

## Why:
I recently fixed an [upstream tree-sitter
issue](https://github.com/tree-sitter/tree-sitter-cpp/pull/239) in the
C++ grammar. Unfortunately, this means we need to take all of the new
`tree-sitter-cpp` changes to pull it in.

## How:
Lots of grammar hacking.

In several places, I augmented the C AST to be more in line with the CPP
AST. This makes it noticeably less simple, but is required, as the
`tree-sitter-c` grammar becomes more complex. This leads to a gross
amount of code duplication between the C and C++ translation code, but
since the typed CSTs are different at the moment, there isn't a good way
to prevent that.

This PR also makes use of the groundwork I laid down in
https://github.com/semgrep/semgrep/pull/9681, which introduces a
`preproc_if_poly` (and friends) types, which are polymorphic types that
encode the structure of the preprocessor statements that occur in C and
C++. Here, we see that we save massive amounts of logic duplication with
minimal boilerplate.

## Test plan:
`make test`

Added a parsing test for the parenthesized assignment thing that all of
this was originally for, too.