Add source information to Pattern_match.t
Now we can actually leverage the source information we are storing in,
e.g., `Target_location.t`. This updates `Pattern_match.t` to have an
associated source and to use it in generating the json output. Some
changes to the Xpattern_match code is required: mostly, a source is
carried along with the file itself. We could consider just using
`Target_location.t` here, but it was unclear to me how much effort is
required.