AST_to_IL: Python: Simplify with-stmts translation
The previous translation was more accurate (yet not 100%) but we really
did not need that accuracy for anything, and it prevented sym-prop to
deduce `x = foo()` from `with foo() as x`. A much simpler translation
that plays well with sym-prop is actually more useful.

Closes CODE-6633

test plan:
make test # new tests