fix(pro): Add kind to Dataflow_svalue.constness type (#9741)
Previously, for any function that the Pro Engine reported returned a
constant value, we would assume that it returned a string constant
value, specifically. This is wrong.

Here, I augment the `Dataflow_svalue.constness` type to include the
specific kind of constant value that a function returns. Then, we use
that kind instead of assuming that it is `Cstr`. Note that this affects
what we expect the hook that the Pro Engine installs to return.

Test plan: Automated tests. No effect expected on this repository.