chore: python: Use the static keyword for static methods in python (#9746)
Simple change to the python parser to lower `@staticmethod` and
`@staticmethod()` to the static keyword. This is used to eliminate some
python specific details in this [semgrep-pro
pr](https://github.com/semgrep/semgrep-proprietary/pull/1298).

# Test Plan
Adds a test to check the behavior of matching. CI should continue to
work with the exception of semgrep pro, which is changed to work in the
PR above.