refactor: Make Path_test_utils.ml
I've encountered another case where it is helpful to create a directory
with a variety of files in it, in order to set up a test. We could mock,
but that's more trouble and this allows for a more authentic integration
test anyway.

This also introduces a new variant of `with_file_tree`:
`with_file_trees`, which can be used in cases where we want multiple
files to exist in the automatically-created workspace directory.

Test plan: Automated tests