fix(release): properly escape single line commands in jsonnet
Recently we switched to using yq for converting jsonnet to yml.

It seems like we need to escape properly for single-line commands.

Release v1.61.0 ran into issues because of this. We were trying
to extract the version number from `refs/tags/v1.61.0`, but because
we didn't escape properly, we ended up getting `tags///tags/v1.61.0`
instead of `v1.61.0`.