fix(release): properly escape single line commands in jsonnet (#9766)
Recently we switched to using yq for converting jsonnet to yml.

It seems like we need to escape properly for single-line commands.

Release v1.61.0 ran into issues because of this. We were trying to
extract the version number from `refs/tags/v1.61.0`, but because we
didn't escape properly, we ended up getting `tags///tags/v1.61.0`
instead of `v1.61.0`.

Searched for all occurrences of backslashes in jsonnet files and
release.yaml was the only file with problematic occurrences, if I'm not
missing anything.

Test plan: actually release 1.61.1 with this change. This change can't
be run in dry run mode unfortunately.