chore(ci): explicitly specify permissions for workflows that need them (#9771)
We've recently been upgraded to Github enterprise, but because of that,
it seems like the default permissions for all workflows are changed.

Some of the checks that were previously passing are now failing. Now we
need to explicitly specify permissions needed (which I think is a good
thing for security).

Looked at all jobs that have the permission field and make all the
depending of those jobs have the least permissions that would allow all
their sub-jobs to be happy.

Test: While some permissions were misconfigured, some checks weren't
running, but now all checks (the # of checks match the checks in
previous PRs before the enterprise upgrade) are running.