tainting: Add `exact: false` sinks
Simplify heuristic to make `exact: true` sinks (the current default) be
backwards compatible with `sink(...)` specs, now we simply require that
the spec formula is an 'And' with `focus`, so it's trivial to know when
the heuristic will trigger.

And add `exact: false` sinks because there are rules that do rely on
those old semantics of sink specs.

test plan:
make test # one new test