fix: Git_wrapper.ls_tree filters paths with spaces
Previously, we used space as a delimiter for the information we get back
from ls-tree. However if a space appeared in a filename we would split
it incorrectly and fail to parse the record, and discard that line as
malformed. This meant we would filter out any information about blobs
whose path had a space.

This commit fixes that issue by replacing space as the separator with a
null byte, which cannot appear in filenames on Linux or MacOS. I think
in theory that such a path could exist on Windows since
    1. WinAPI strings generally explicitly store length
    2. AFAIK Windows uses UTF-16, so a null byte could occur as part of
       a 2-byte sequence.

That said, please do not null byte in a file path. So, this should
generally be fine (and is certainly an improvement).