test: allow for a wider range of paths for `--dump-engine-path` test
In nixpkgs semgrep-core is being output to a path such as:
`/nix/store/k8r0326l6ms2v0dhgj706r5g7f4ha_i8y-semgrep-core-1.61.1/bin/semgrep-core`

If that is output by `semgrep` when used with `--dump-engine-path` the test fails.
This change expands the regex to allow for that match too.