tainting: Add heuristic to choose label to report trace
When using taint labels, not all labels are equally relevant for the
finding, but Semgrep can only report one label at present (until we
generalize taint traces). Instead of picking always an arbitrary label,
we now look into the 'requires' of the sink, and if it has the shape
`A and ...` then we pick 'A' as our preferred label to report the trace.

Closes FLOW-65

test plan:
Run Semgrep with `--dataflow-traces` on https://semgrep.dev/playground/s/AbDeL
  #^ Now you get a taint trace for the label 'USER_INPUT'.