fix: Add `--` to git diff command (#9775)
## Problem

Some repositories contain a file in the root dir with the same name as
that of the main branch. When this is the case, running a git diff
against the main branch results in an error.

```
❯ ls
main
[...]

❯ git symbolic-ref HEAD --short
main

❯ git diff main
fatal: ambiguous argument 'main': both revision and filename
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'
```

## Solution

Append a `--` to the command as recommended by git. This clarifies we're
providing the name of a _branch_ to diff, not the name of a file.

```
❯ git diff main --
```

Adding this parameter has no effect for repositories without a root file
with the same name as the main branch, so we can safely add it to our
command across the board.