refactor(rules): decouple conditions from `And`s (#9696)
## What:
This PR changes the internal representation of rules to be such that we
no longer solely associate conditions like `metavariable-regex` and
friends with `all`/`patterns` operators.

## Why:
In the new syntax, we permit `where` to be added to patterns in
arbitrary places. Currently, we compile all instances of `where` into a
vestigial `And`, which is rather redundant. It would be nicer if our
internal representation lined up more neatly with the actual structure
of rules.

## How:
This PR splits up the `Rule.t` type into the `Rule.rule_kind` type,
which is a sum of the previous `Rule.t` variants, and `Rule.t`, which
combines a `Rule.rule_kind` with additional information (notably,
conditions and focus).

## Test plan:
All tests still pass.