feat(matching): allow interpolated vars to work with metavariable regex (#9808)
## What:
This PR makes it so `metavariable-regex` on an interpolated string can
actually match based off of the variables being interpolated.

## Why:
We want to be able to, in particular, match patterns like
`"${var.varname}"` in Terraform, via binding a metavariable to it and
checking the contents with `metavariable-regex`.

## How:
We can't do this naively, because `var.varname` is not an `N`
metavariable. This means the propagated value is contained inside of it,
but `Eval_generic` doesn't know how to access it.

To make `Eval_generic` aware of it, we added logic to simulate the
interpolation of the string. This change is not Terraform-specific.

## Test plan:
Automated tests.

Closes SAF-865