test: Make autofix e2e tests more extensive
The most interesting changes here are in `test_autofix.py`. The rest are
changes to the snapshots.

Previously, we tested only the JSON output. We also did not test for
what appears on stderr. I'm not sure why, but we do print a fair amount
of information intended for the end user to consume on stderr rather
than stdout (see the snapshots). This seems important to preserve.

This change adds tests for text output in addition to JSON output for
all of our existing autofix e2e tests. It also includes stderr in the
snapshot. Unfortunately, this means that this test needs to be marked
`osemfail` because osemgrep puts a fair amount of logspew on stderr (and
perhaps other issues, but that is the most obvious).

Test plan: This is a test