osemgrep: apply autofixes in the findings order (#9868)
The following test was failing in another, [unrelated
branch](https://github.com/semgrep/semgrep/pull/9624):
```
PYTEST_LOG_LEVEL=debug PYTEST_USE_OSEMGREP=true pipenv run pytest -k test_autofix[rules/autofix/exact-collision.yaml-autofix/collision.py-not-dryrun]
```
It's not clear what happened in that branch but I found out that the
order in which autofixes are applied was different in osemgrep than in
pysemgrep. I fixed it and added a test to ensure they use the same
order.

This also revealed that the results are not printed in the same order as
the original findings that determines the order of the autofixes.
There's also a problem in pysemgrep with some fixes being reported as
applied even though they were not applied due to overlaps. I left this
as a todo.