feat: interproc taint scans run by default for logged in users
Previously, we required the Pro engine toggle to be on to run interproc taint
(PRO_INTRAFILE) scans. Now that we've validated that these scans give good
improvement in results with virtually no perf regression, and that we have the
pro binary in the docker image, it's time to give users more powerful analysis
by default.

Test plan: automated unit tests, canary image