feat: allow triggering historical scan from CI (#9870)
This adds the required glue code to run an additional scan with the
required flags in the instance where historical scanning is enabled via
the app and the secrets product is enabled.

Also fixes a minor issue where having only secrets enabled with no rules
would not result in a warning we displayed in the case where only sast
was enabled but no rules were sent.

Test plan:  Manually tested with the test from
first sub-commit (see github; removed in the squash)
with local copy of the pro binary. Test to be added to pro once we
have equivalent testing capabilities: either repo merge or if it is
made possible by various py->osemgrep migration work.