Use 'make test' to run pytest tests in CI (#9879)
This PR reduces discrepancies between what's used in CI and the old
`make test` command that runs pytest tests.
* the default `make test` command in `/cli` is now the reference and it
is used by CI jobs. It used to be `make test-for-ci`.
* the old `make test` is now `make test-dev`. It's for developers who
want tests to fail faster. The main issue that prevents us from merging
the two is that not all the tests will run if a test fails in the
"quick" or in the "kinda_slow" groups. In CI, we want all the tests to
run and get a complete recap at the end. See todo comments in the
makefile.
* the old pytest command used in CI was selecting fewer tests (1567 vs.
1728) than when explicitly passing a folder as an argument. ~~We don't
know why.~~ This was fixed by configuring pytest to run by default on
`/cli/tests/default` instead of some selection of folders. This moved a
lot of files from `/cli/tests/` to `/cli/tests/default/`.
* the old pytest command used in CI wasn't checking that all the tests
had a speed marker. This is fixed.

Various tests were using paths of the form `TESTS_PATH / "e2e" /
"targets"` and I had to update it everywhere. Now it uses a variable
`TARGETS_PATH` instead. Same with rules and `RULES_PATH`.