feat: replace libpcre with libpcre2
This replaces relevant occurances, mostly as exposed through Pcre_, of
Pcre with Pcre2 (via Pcre2_) and updates the build to support libpcre2
in addition to libpcre.

The former is used in general, and the latter is specifically for
generic mode, until we either update the generate regex, or guarantee a
sufficiently high version of PCRE2 as to not cause the test failures
Martin saw with 10.34 (e.g., 10.43 or later).

It looks like there are many places we used Pcre instead of Pcre_ which
would be good to clean up and just use Pcre_ instead.

This also:
- Removes `Pcre_.string_of_error` since we already have a derived show
  implementation which is the same.
- Adds `Pcre_.quote`. Previously we used the version directly from the
  Pcre library; we should wrap this for abstraction since we do the same
  for other library functionality.