Initial replacement of Pcre with Pcre2
This simply replaces relevant occurances, as exposed through Pcre_, of
Pcre with Pcre2. It does not yet attempt to ensure static linkage or
address the build system generally.

It looks like there are many places we used Pcre instead of Pcre_ which
would be good to clean up and just use Pcre_ instead. Currently I am
leaning toward renaming that module too, given it is easy to
"accidentally" use the library directly, rather than through a wrapper
(and a wrapper makes it easier to do migrations like these in the
future).