refactor: Refactor how osemgrep creates `fixed_lines` (#9921)
The way this was implemented was confusing and not reusable.

This is mostly a refactor, though I did also address some behaviors in
osemgrep that were inconsistent with pysemgrep. But, I don't think it's
worth a changelog entry since it only affects osemgrep.

Test plan:
- Unit tests
- I'm using this utility in my in-progress work to replace the Python
autofix application code with an RPC call. It currently passes all
autofix e2e tests.